+++ app/src/app/enums/mime-types.enum.ts
@@ -1,0 +1,7 @@
+export enum MimeTypes {
+    video = 'video/webm',
+    videoMp4 = 'video/mp4',
+    audioWebm = 'audio/webm;codecs=opus',
+    audioMp4 = 'audio/mp4;codecs=mp4a.67',
+    audioMp3 = 'audio/mpeg',
+}