+++ app/src/app/enums/recorder-state.enum.ts
@@ -1,0 +1,5 @@
+export enum RecorderState {
+    inactive = 'inactive',
+    recording = 'recording',
+    paused = 'paused'
+}