+++ app/src/app/interfaces/environment.interface.ts
@@ -1,4 +1,5 @@
 export interface Enviroment {
+    apiUrl: string;
     production: boolean;
     version: string;
 }