+++ app/src/app/models/animator.ts
@@ -4,19 +4,24 @@ import { Platform } from '@ionic/angular';
 import { BaseService } from '@services/base/base.service';
 import { BehaviorSubject, Observable } from 'rxjs';
 import { first } from 'rxjs/operators';
+import { saveAs } from 'file-saver';
 import * as WebMWriter from 'webm-writer';
+import * as zip from '@zip.js/zip.js';
+import { MimeTypes } from '@enums/mime-types.enum';
+import { RecorderState } from '@enums/recorder-state.enum';
+import { ImagesService } from '@services/images/images.service';
 // eslint-disable-next-line @typescript-eslint/naming-convention
-declare const WebPEncoder: any;
 declare const webm: any;
 @Injectable({
     providedIn: 'root'
 })
 export class Animator {
-    audio: any;
+    audio: HTMLAudioElement;
     audioBlob: any;
     audioChunks: any[];
+    audioMimeType: string;
     audioRecorder: MediaRecorder;
-    exported: any;
+    audioStream: MediaStream;
     rotated: boolean;
     frames: any[];
     framesInFlight: number;
@@ -24,8 +29,6 @@ export class Animator {
     height: number;
     isStreaming: boolean;
     isRecording: boolean;
-    loadFinishPending: boolean;
-    messageDiv: any;
     name: any;
     playCanvas: any;
     playContext: any;
@@ -34,7 +37,7 @@ export class Animator {
     snapshotContext: any;
     video: any;
     videoSourceId: any;
-    videoStream: any;
+    videoStream: MediaStream;
     width: any;
     zeroPlayTime: number;
 
@@ -44,10 +47,11 @@ export class Animator {
     constructor(
         // TODO if possible get rid of injectable again
         public baseService: BaseService,
-        private platform: Platform
+        private platform: Platform,
+        private imagesService: ImagesService,
     ) {
         this.isAnimatorPlaying = new BehaviorSubject(false);
-        this.frameRate = new BehaviorSubject(12.0);
+        this.frameRate = new BehaviorSubject(6.0);
     }
 
     getIsPlaying(): Observable<any> {
@@ -61,26 +65,30 @@ export class Animator {
     /*
     * Init method is used to initialize properties
     */
-    public init(video, snapshotCanvas, playCanvas, layoutOptions): void {
-        this.audio = null;
-        this.audioBlob = null;
-        this.audioChunks = [];
-        this.audioRecorder = null;
-        this.frames = [];
-        this.framesInFlight = 0;
-        this.frameWebps = [];
-        this.isStreaming = true;
-        this.loadFinishPending = false;
-        this.name = null;
-        this.playCanvas = playCanvas;
-        this.playContext = playCanvas.getContext('2d');
-        this.playTimer = null;
-        this.rotated = false;
-        this.snapshotCanvas = snapshotCanvas;
-        this.snapshotContext = snapshotCanvas.getContext('2d');
-        this.video = video;
-        this.videoStream = null;
-        this.zeroPlayTime = 0;
+    public async init(video: any, snapshotCanvas: any, playCanvas: any, layoutOptions: any, hasData?: boolean): Promise<void> {
+        if (!hasData) {
+            this.audio = null;
+            this.audioBlob = null;
+            this.audioChunks = [];
+            // determine if os is iOS or browser is safari, if so use other codec to store audio
+            this.audioMimeType = this.getAudioMimeType();
+            this.audioStream = await navigator.mediaDevices.getUserMedia({ audio: true });
+            this.frames = [];
+            this.framesInFlight = 0;
+            this.frameWebps = [];
+            this.isStreaming = true;
+            this.name = null;
+            this.playCanvas = playCanvas;
+            this.playContext = playCanvas.getContext('2d');
+            this.playTimer = null;
+            this.rotated = false;
+            this.snapshotCanvas = snapshotCanvas;
+            this.snapshotContext = snapshotCanvas.getContext('2d');
+            this.video = video;
+            this.videoStream = null;
+            this.zeroPlayTime = 0;
+        }
+        console.log('🚀 ~ file: animator.ts ~ line 334 ~ Animator ~ returnnewPromise ~ this.audioMimeType', this.audioMimeType);
 
         this.setDimensions(layoutOptions);
     }
@@ -89,18 +97,18 @@ export class Animator {
     * Method is used to attach a media stream to the video component
     */
     public async attachStream(sourceId: any, layoutOptions: LayoutOptions, facingMode?: string): Promise<any> {
-        console.log('🚀 ~ file: animator.ts ~ line 92 ~ Animator ~ attachStream ~ layoutOptions', layoutOptions);
+        // console.log('🚀 ~ file: animator.ts ~ line 92 ~ Animator ~ attachStream ~ layoutOptions', layoutOptions);
         const constraints = {
             audio: false,
-            frameRate: 15,
+            frameRate: 30,
             video: null
         };
 
         facingMode = facingMode ? facingMode : 'user';
+        const aspectRatio = (layoutOptions.isPortrait) ? layoutOptions.height / layoutOptions.width
+            : layoutOptions.width / layoutOptions.height;
 
         if (this.platform.is('ios') || this.platform.is('android')) {
-            const aspectRatio = (layoutOptions.isPortrait) ? layoutOptions.height / layoutOptions.width
-            : layoutOptions.width / layoutOptions.height;
             constraints.video = {
                 // strange bug - width and height needs to be swaped
                 width: this.platform.is('android') ? layoutOptions.height : layoutOptions.width,
@@ -113,7 +121,8 @@ export class Animator {
                 constraints.video = {
                     width: layoutOptions.width,
                     height: layoutOptions.height,
-                    sourceId
+                    sourceId,
+                    aspectRatio
                 };
             } else {
                 constraints.video = true;
@@ -125,6 +134,7 @@ export class Animator {
             this.video.srcObject = stream;
             this.videoStream = stream;
             this.isStreaming = true;
+
             return stream;
         } catch (err) {
             console.error(err);
@@ -149,7 +159,7 @@ export class Animator {
         context.drawImage(this.video, 0, 0, this.width, this.height);
         this.frames.push(imageCanvas);
 
-        const promise = new Promise(((resolve, reject) => {
+        const promise = await new Promise(((resolve, reject) => {
             this.snapshotContext.clearRect(0, 0, this.width, this.height);
             this.snapshotContext.drawImage(imageCanvas, 0, 0, this.width, this.height);
             imageCanvas.toBlob(blob => { resolve(blob); }, 'image/webp');
@@ -187,6 +197,9 @@ export class Animator {
         this.name = null;
     }
 
+    /*
+   * Method is used to toggle camera
+   */
     public async toggleCamera(layoutOptions: LayoutOptions) {
         if (this.video.paused) {
             if (this.video.srcObject && this.video.srcObject.active) {
@@ -235,29 +248,37 @@ export class Animator {
         this.setAudioSrc(null);
     }
 
-    startPlay(noAudio) {
-        return new Promise((resolve, reject) => {
-            if (!this.frames.length) {
-                resolve(false);
-                return;
-            }
-            this.snapshotCanvas.style.visibility = 'hidden';
-            this.video.pause();
-            this.drawFrame(0, this.playContext);
-            this.zeroPlayTime = performance.now();
-            this.playTimer = setTimeout(this.playFrame.bind(this), this.frameTimeout(), 1, resolve);
-            if (this.audio && !noAudio) {
+    public async startPlay(noAudio: boolean) {
+        if (!this.frames.length) {
+            return;
+        }
+        this.snapshotCanvas.style.visibility = 'hidden';
+        this.video.pause();
+        this.drawFrame(0, this.playContext);
+        this.zeroPlayTime = performance.now();
+        this.playTimer = setTimeout(this.playFrame.bind(this), this.frameTimeout(), 1);
+        await this.playAudio(noAudio);
+        this.isAnimatorPlaying.next(true);
+    }
+
+    async playAudio(noAudio: boolean) {
+        if (this.audio && !noAudio) {
+            try {
                 this.audio.currentTime = 0;
-                this.audio.play();
+                await this.audio.play();
+            } catch (error: any) {
+                this.baseService.toastService.presentToast({
+                    message: this.baseService.translate.instant('toast_animator_audio_play_error'),
+                    color: 'danger',
+                });
             }
-            this.isAnimatorPlaying.next(true);
-        });
+        }
     }
 
     endPlay(cb) {
         if (this.isPlaying()) { clearTimeout(this.playTimer); }
         this.playTimer = null;
-        if (this.audioRecorder && this.audioRecorder.state === 'recording') {
+        if (this.audioRecorder && this.audioRecorder.state === RecorderState.recording) {
             this.audioRecorder.stop();
         } else if (this.audio) {
             this.audio.pause();
@@ -269,27 +290,10 @@ export class Animator {
         if (cb) { cb(); }
     }
 
-    playFrame(frameNumber, cb) {
-        if (frameNumber >= this.frames.length) {
-            this.playTimer = setTimeout(this.endPlay.bind(this), 1000, cb);
-        } else {
-            this.drawFrame(frameNumber, this.playContext);
-            const timeout = this.zeroPlayTime + ((frameNumber + 1) * this.frameTimeout()) - performance.now();
-            this.playTimer = setTimeout(this.playFrame.bind(this), timeout, frameNumber + 1, cb);
-        }
-        // console.log('🚀 ~ file: animator.ts ~ line 245 ~ Animator ~ playFrame ~  this.playTimer', this.playTimer);
-    }
-
-    drawFrame(frameNumber, context) {
-        context.clearRect(0, 0, this.width, this.height);
-        context.drawImage(this.frames[frameNumber], 0, 0, this.width, this.height);
-    }
-
-    frameTimeout() {
-        return 1000.0 / this.frameRate.getValue();
-    }
-
-    detachStream() {
+    /*
+    * Method is used to detach current media stream in case of camera change or data is cleared by the user
+    */
+    public detachStream() {
         if (!this.video.srcObject) {
             return;
         }
@@ -303,15 +307,91 @@ export class Animator {
         return !!this.playTimer;
     }
 
-    loadFinished() {
-        if (this.frames.length) {
+    /*
+    * recordAudio method is used to record audio using browser MediaRecorder API
+    */
+    public async recordAudio(): Promise<Blob> {
+        if (!this.frames.length) {
+            return;
+        }
+        const state = this.audioRecorder ? this.audioRecorder.state : RecorderState.inactive;
+        if (state === RecorderState.recording) {
+            return;
+        }
+
+        return new Promise((async (resolve, reject) => {
+            this.audioRecorder = new MediaRecorder(this.audioStream, { mimeType: this.audioMimeType });
+            this.audioRecorder.ondataavailable = ((event: any) => {
+                console.log('🚀 ~ file: animator.ts ~ line 318 ~ Animator ~ returnnewPromise ~ event', event);
+                this.audioChunks.push(event.data);
+            });
+            this.audioRecorder.onstop = ((evt: any) => {
+                this.audioRecorder = null;
+                const blob = new Blob(this.audioChunks, { type: this.audioMimeType });
+                this.audioChunks = [];
+                resolve(blob);
+            });
+            // pass true to do not play audio at the same time
+            this.startPlay(true);
+            this.audioRecorder.start();
+        }));
+    }
+
+    /*
+    * setDimensions method is used to set dimension width and height of components
+    */
+    public setDimensions(layoutOptions: LayoutOptions): void {
+        // console.log('🚀 ~ file: animator.ts ~ line 403 ~ Animator ~ setDimensions ~ layoutOptions', layoutOptions);
+        this.width = layoutOptions.width;
+        this.height = layoutOptions.height;
+        this.video.width = this.width;
+        this.video.height = this.height;
+        this.snapshotCanvas.width = this.width;
+        this.snapshotCanvas.height = this.height;
+        this.playCanvas.width = this.width;
+        this.playCanvas.height = this.height;
+    }
+
+    /*
+   * Method is used to trigger file loading process
+   */
+    public async load(file: any): Promise<any> {
+        const result = await this.readFile(file);
+        console.log('🚀 ~ file: animator.ts ~ line 359 ~ Animator ~ load ~ result', result);
+        try {
+            await this.decodeFile(await new Response(result[0]).arrayBuffer());
+            console.log('after decode');
+            if (result[1]) {
+                console.log('🚀 ~ file: animator.ts ~ line 363 ~ Animator ~ load ~ result[1]', result[1]);
+                this.setAudioSrc(result[1], MimeTypes.audioMp3);
+            }
             this.snapshotContext.clearRect(0, 0, this.width, this.height);
             this.snapshotContext.drawImage(this.frames[this.frames.length - 1], 0, 0, this.width, this.height);
-            this.startPlay(null);
+            return;
+        } catch (err) {
+            console.log('🚀 ~ file: animator.ts ~ line 370 ~ Animator ~ load ~ err', err);
+            return;
         }
     }
 
-    setAudioSrc(blob) {
+    /*
+    * Method is used to trigger file saving as draft process
+    */
+    public async saveDraft(filename: string) {
+        const videoBlob = await this.createVideoBlob();
+        const audioBlob = (this.audio) ? this.audioBlob : null;
+        console.log('🚀 ~ file: animator.ts ~ line 378 ~ Animator ~ save ~ audioBlob', audioBlob);
+        const dataURI = await this.createZipFile(videoBlob, audioBlob);
+        saveAs(dataURI, filename + '.zip', { autoBom: true });
+        URL.revokeObjectURL(dataURI);
+        return;
+    }
+
+    /*
+    * Method is used to set audio source from blob
+    */
+    public setAudioSrc(blob: Blob, mimeType?: MimeTypes) {
+        console.log('🚀 ~ file: animator.ts ~ line 513 ~ Animator ~ setAudioSrc ~ blob', blob, mimeType);
         this.audioBlob = blob;
         if (this.audio) {
             URL.revokeObjectURL(this.audio.src);
@@ -319,13 +399,43 @@ export class Animator {
         }
         if (blob) {
             this.audio = document.createElement('audio');
-            this.audio.src = URL.createObjectURL(blob);
+            const sourceElement = document.createElement('source');
+            this.audio.appendChild(sourceElement);
+            sourceElement.src = URL.createObjectURL(blob);
+            console.log('🚀 ~ file: animator.ts ~ line 527 ~ Animator ~ setAudioSrc ~ URL.createObjectURL(blob)', URL.createObjectURL(blob));
+            sourceElement.type = (mimeType) ? mimeType : this.getAudioMimeType();
+            this.audio.load();
         }
     }
 
-    public async save(filename) {
-        filename = filename || 'StopMotion';
-        if (!filename.endsWith('.webm')) { filename += '.webm'; }
+    /*
+    * Method is used to play single frame and apply timeout for next frame and update playTimer
+    */
+    private playFrame(frameNumber: number, cb: () => void) {
+        console.log('🚀 ~ file: animator.ts ~ line 382 ~ Animator ~ playFrame ~ frameNumber', frameNumber, this.frames.length);
+        if (frameNumber >= this.frames.length) {
+            // this.playTimer = setTimeout(this.endPlay.bind(this), 1000, cb);
+            this.endPlay(cb);
+        } else {
+            this.drawFrame(frameNumber, this.playContext);
+            const timeout = this.zeroPlayTime + ((frameNumber + 1) * this.frameTimeout()) - performance.now();
+            this.playTimer = setTimeout(this.playFrame.bind(this), timeout, frameNumber + 1, cb);
+        }
+    }
+
+    /*
+   * Method is used to clear canvas and draw current frame
+   */
+    private drawFrame(frameNumber, context) {
+        context.clearRect(0, 0, this.width, this.height);
+        context.drawImage(this.frames[frameNumber], 0, 0, this.width, this.height);
+    }
+
+    /*
+    * Method is used to create video blob
+    * Depending on platform differnt types of enconding are used
+    */
+    private async createVideoBlob(): Promise<Blob> {
         const frameRate = await this.getFramerate().pipe(first()).toPromise();
         const videoWriter = new WebMWriter({
             quality: 0.95,    // WebM image quality from 0.0 (worst) to 0.99999 (best), 1.00 (VP8L lossless) is not supported
@@ -339,82 +449,108 @@ export class Animator {
             // If not specified this defaults to the same value as `quality`.
         });
 
-        for (const frame of this.frames) {
-            if (this.isSafari()) {
-                const encoder = new WebPEncoder();
-                const config = {
-                    // eslint-disable-next-line @typescript-eslint/naming-convention
-                    extra_info_type: 0
-                };
-                encoder.WebPEncodeConfig(config); //when you set the config you must it do for every WebPEncode... new
-                const out = { output: '' };
-                //w*4 desc: w = width, 3:RGB/BGR, 4:RGBA/BGRA
-                const ctx = frame.getContext('2d').getImageData(0, 0, this.width, this.height);
-                encoder.WebPEncodeRGBA(ctx.data, this.width, this.height, this.width * 4, 75, out);
-                const base64URI = `data:image/webp;base64,${btoa(out.output)}`;
-                videoWriter.addFrame(base64URI);
-            } else {
+        if (this.isSafari()) {
+            for (const frame of this.frameWebps) {
+                const image = await this.imagesService.convertImages(frame);
+                videoWriter.addFrame(this.uint8ToBase64(image));
+            }
+        } else {
+            for (const frame of this.frames) {
                 videoWriter.addFrame(frame);
             }
         }
 
         const blob = await videoWriter.complete();
-        this.exported = blob;
-        const url = URL.createObjectURL(blob);
-        const downloadLink = document.createElement('a');
-        downloadLink.download = filename;
-        downloadLink.href = url;
-        downloadLink.click();
-        URL.revokeObjectURL(url);
         return blob;
     }
 
-    public async recordAudio(stream) {
-        // determine if os is iOS or browser is safari, if so use other codec to store audio
-        const mimeType = this.isSafari() ? 'audio/mp4' : 'audio/webm;codecs=opus';
-        // const mimeType = 'video/webm';
-        return new Promise(((resolve, reject) => {
-            if (!this.frames.length) {
-                resolve(null);
-                return;
-            }
-            const state = this.audioRecorder ? this.audioRecorder.state : 'inactive';
-            if (state === 'recording') {
-                resolve(null);
-                return;
-            }
-            this.audioRecorder = new MediaRecorder(stream, { mimeType });
-            this.audioRecorder.ondataavailable = (evt => {
-                this.audioChunks.push(evt.data);
-            }).bind(this);
-            this.audioRecorder.onstop = (evt => {
-                this.audioRecorder = null;
-                this.setAudioSrc(new Blob(this.audioChunks, { type: mimeType }));
-                this.audioChunks = [];
-                resolve(this.audioBlob);
-            }).bind(this);
-            this.startPlay(true);
-            this.audioRecorder.start();
-        }));
+    /*
+    * Method is used to create zip file of video and audio (if available)
+    */
+    private async createZipFile(videoBlob: Blob, audioBlob: Blob): Promise<string> {
+        zip.configure({ useWebWorkers: false });
+        const zipWriter = new zip.ZipWriter(new zip.Data64URIWriter('application/zip'));
+        await zipWriter.add('video.webm', new zip.BlobReader(videoBlob));
+        if (audioBlob) {
+            await zipWriter.add('audio.webm', new zip.BlobReader(audioBlob));
+        }
+        const dataURI = await zipWriter.close();
+        return dataURI;
     }
 
     /*
-    * setDimensions method is used to set dimension width and height of components
+    * Method is used to decode array buffer to single frames, export framerate
     */
-    public setDimensions(layoutOptions: LayoutOptions): void {
-        console.log('🚀 ~ file: animator.ts ~ line 403 ~ Animator ~ setDimensions ~ layoutOptions', layoutOptions);
-        this.width = layoutOptions.width;
-        this.height = layoutOptions.height;
-        this.video.width = this.width;
-        this.video.height = this.height;
-        this.snapshotCanvas.width = this.width;
-        this.snapshotCanvas.height = this.height;
-        this.playCanvas.width = this.width;
-        this.playCanvas.height = this.height;
+    private async decodeFile(fileBuffer: ArrayBuffer) {
+        const animator = this;
+        const result = await new Promise((resolve) => {
+            webm.decode(fileBuffer,
+                (width: number, height: number) => {
+                    this.setDimensions({
+                        width: this.width,
+                        height: this.height
+                    } as any);
+                },
+                (frameRate: number) => {
+                    this.setFramerate(Math.round(frameRate));
+                },
+                animator.addFrameVP8.bind(animator, this.frames.length, resolve),
+                animator.setAudioSrc.bind(animator));
+        });
+        console.log('🚀 ~ file: animator.ts ~ line 507 ~ Animator ~ result ~ result', result);
+        return result;
     }
 
+    /*
+    * Method is used to read entire zip file
+    */
+    private async readFile(file: any): Promise<Blob[]> {
+        const reader = new zip.ZipReader(new zip.BlobReader(file));
+        const entries = await reader.getEntries();
+        const blobs = [];
+        await Promise.all(entries.map(async (entry: any, index: number) => {
+            console.log('🚀 ~ file: animator.ts ~ line 496 ~ Animator ~ awaitPromise.all ~ entry', entry);
+            const blob = await this.readFileEntry(entry, index);
+            console.log('🚀 ~ file: animator.ts ~ line 498 ~ Animator ~ awaitPromise.all ~ blob', blob);
+            blobs.push(blob);
+        }));
+        await reader.close();
+        if (blobs.length > 1 && blobs[0].type !== MimeTypes.video) {
+            // swap array elements if audio is first
+            blobs.unshift(blobs.pop());
+        }
+        console.log('🚀 ~ file: animator.ts ~ line 505 ~ Animator ~ readFile ~ blobs', blobs);
+        console.log('🚀 ~ file: animator.ts ~ line 531 ~ Animator ~ readFile ~ this.frameWebps', this.frameWebps);
+        console.log('🚀 ~ file: animator.ts ~ line 531 ~ Animator ~ readFile ~ this.frames', this.frames);
+        return blobs;
+    }
 
-    addFrameVP8(frameOffset, callback, blob, idx) {
+    /*
+    * Method is used to read file inside of zip file
+    */
+    private async readFileEntry(entry: any, index: number): Promise<Blob> {
+        const type = (index === 0) ? MimeTypes.video : MimeTypes.audioMp3;
+        return await entry.getData(new zip.BlobWriter(type));
+    }
+
+    /*
+    * Method is used to determine wether browser has user agent safari or is used on ios
+    */
+    private isSafari(): boolean {
+        return (navigator.userAgent.indexOf('Safari') !== -1 && navigator.userAgent.indexOf('Chrome') === -1) || this.platform.is('ios');
+    }
+
+    /*
+    * Method is used to get proper audio mime type
+    */
+    private getAudioMimeType(): string {
+        return this.isSafari() ? MimeTypes.audioMp4 : MimeTypes.audioWebm;
+    }
+
+    /*
+    * Method is used to add single frames from files after it is loaded
+    */
+    private addFrameVP8(frameOffset: number, callback: any, blob: Blob, index: number) {
         let blobURL = URL.createObjectURL(blob);
         const image = new Image(this.width, this.height);
         this.framesInFlight++;
@@ -435,71 +571,42 @@ export class Animator {
                 image.src = blobURL;
             }
         });
-        // .bind(this));
 
-        image.addEventListener('load', (evt: any) => {
+        image.addEventListener('load', async (evt: any) => {
             const newCanvas = document.createElement('canvas');
             newCanvas.width = this.width;
             newCanvas.height = this.height;
             newCanvas.getContext('2d', { alpha: false }).drawImage(evt.target, 0, 0, this.width, this.height);
-            this.frames[frameOffset + idx] = newCanvas;
-            this.frameWebps[frameOffset + idx] = new Promise((resolve, reject) => {
+            this.frames[frameOffset + index] = newCanvas;
+            this.frameWebps[frameOffset + index] = await new Promise((resolve, reject) => {
                 resolve(blob);
             });
             this.framesInFlight--;
             URL.revokeObjectURL(blobURL);
             if (this.framesInFlight === 0) { callback(); }
         });
-        // .bind(this));
 
         image.src = blobURL;
     }
 
-    public async load(file: any): Promise<any> {
-        const result = await this.readFile(file);
-        try {
-            await this.decodeFile(result);
-            this.loadFinished();
-            return;
-        } catch (err) {
-            console.log('🚀 ~ file: animator.ts ~ line 422 ~ Animator ~ load ~ err', err);
-            return;
-        }
-    }
-
-    private async decodeFile(fileBuffer: ArrayBuffer) {
-        const animator = this;
-        return new Promise((resolve) => {
-            webm.decode(fileBuffer,
-                // animator.setDimensions.bind(animator),
-                (width: number, height: number) => {
-                    this.setDimensions({
-                        width: this.width,
-                        height: this.height
-                    } as any);
-                },
-                // frameRateCB,
-                (frameRate: number) => {
-                    console.log('🚀 ~ file: animator.ts ~ line 400 ~ Animator ~ load ~ frameRate', frameRate);
-                    this.setFramerate(Math.round(frameRate));
-                },
-                animator.addFrameVP8.bind(animator, this.frames.length, resolve),
-                animator.setAudioSrc.bind(animator));
-        });
-    }
-
-    private async readFile(file: any): Promise<ArrayBuffer> {
-        return new Promise((resolve) => {
-            const reader = new FileReader();
-            reader.onloadend = (event: any) => resolve(event.target.result);
-            reader.readAsArrayBuffer(file);
-        });
+    /*
+    * Method is used to calculate timeout between frames
+    */
+    private frameTimeout() {
+        return 1000.0 / this.frameRate.getValue();
     }
 
     /*
-    * Method is used to determine wether browser has user agent safari or is used on ios
+    * Method is used to convert ArrayBuffer to base64 string
     */
-    private isSafari(): boolean {
-        return (navigator.userAgent.indexOf('Safari') !== -1 && navigator.userAgent.indexOf('Chrome') === -1) || this.platform.is('ios');
+    private uint8ToBase64(buffer: ArrayBuffer): string {
+        let binary = '';
+        const bytes = new Uint8Array(buffer);
+        const len = bytes.byteLength;
+        for (let i = 0; i < len; i++) {
+            binary += String.fromCharCode(bytes[i]);
+        }
+        return `data:image/webp;base64,${window.btoa(binary)}`;
     }
 }
+