+++ app/src/app/pages/animator/components/undo-button/undo-button.component.ts
@@ -12,7 +12,8 @@ export class UndoButtonComponent {
     private animatorService: AnimatorService
   ) { }
 
-  onClick() {
+  public onClick($event: Event): void {
+    $event.preventDefault();
     this.animatorService.undoCapture();
   }