+++ app/src/app/pages/animator/guards/animator.guard.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { AnimatorGuard } from './animator.guard';
+
+describe('AnimatorGuard', () => {
+    let service: AnimatorGuard;
+
+    beforeEach(() => {
+        TestBed.configureTestingModule({});
+        service = TestBed.inject(AnimatorGuard);
+    });
+
+    it('should be created', () => {
+        expect(service).toBeTruthy();
+    });
+});