+++ app/src/environments/environment.prod.ts
@@ -1,4 +1,5 @@
 export const environment = {
+  apiUrl: 'https://stop-motion.web-punks.com/api',
   production: true,
-  version: '1.0.0'
+  version: '1.1.0'
 };