+++ app/src/environments/environment.ts
@@ -5,8 +5,9 @@
 import { Enviroment } from '@interfaces/environment.interface';
 
 export const environment: Enviroment = {
+  apiUrl: 'http://localhost:9000',
   production: false,
-  version: '1.0.0'
+  version: '1.1.0'
 };
 
 /*