+++ app/CHANGELOG.md
@@ -12,3 +12,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 ### Fixed
+
+## [1.1.0] - 2022-06-21
+
+### Added
+
+- Added logic to store and download audio/video as zip
+- Added Docker Setup
+- Added guard and alert for animator page
+
+## [1.0.0] - 2022-02-02
+
+- Initial Release